Remove DefaultNetworkListener and add default implementation directly in the interface
