+++ iidm/iidm-impl/src/test/java/com/powsybl/iidm/network/impl/OperationalLimitsGroupImplTest.java
@@ -26,7 +26,7 @@ class OperationalLimitsGroupImplTest {
     void customOperationalLimitsGroupTest() {
         Network network = EurostagTutorialExample1Factory.create();
         boolean[] updated = new boolean[1];
-        network.addListener(new DefaultNetworkListener() {
+        network.addListener(new NetworkListener() {
             @Override
             public void onUpdate(Identifiable<?> identifiable, String attribute, String variantId, Object oldValue, Object newValue) {
                 assertEquals("NHV1", identifiable.getId());
