+++ iidm/iidm-modification/src/test/java/com/powsybl/iidm/modification/topology/RemoveFeederBayTest.java
@@ -46,7 +46,7 @@ class RemoveFeederBayTest {
     }
 
     private void addListener(Network network) {
-        network.addListener(new DefaultNetworkListener() {
+        network.addListener(new NetworkListener() {
             @Override
             public void beforeRemoval(Identifiable id) {
                 beforeRemovalObjects.add(id.getId());
