+++ iidm/iidm-modification/src/test/java/com/powsybl/iidm/modification/topology/RemoveSubstationTest.java
@@ -14,8 +14,8 @@ import com.powsybl.commons.report.ReportNode;
 import com.powsybl.iidm.modification.AbstractNetworkModification;
 import com.powsybl.iidm.modification.NetworkModification;
 import com.powsybl.iidm.modification.NetworkModificationImpact;
-import com.powsybl.iidm.network.DefaultNetworkListener;
 import com.powsybl.iidm.network.Network;
+import com.powsybl.iidm.network.NetworkListener;
 import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
@@ -38,7 +38,7 @@ class RemoveSubstationTest extends AbstractModificationTest {
     }
 
     private void addListener(Network network) {
-        network.addListener(new DefaultNetworkListener() {
+        network.addListener(new NetworkListener() {
             @Override
             public void afterRemoval(String id) {
                 removedObjects.add(id);
