+++ iidm/iidm-modification/src/test/java/com/powsybl/iidm/modification/topology/RemoveVoltageLevelTest.java
@@ -41,7 +41,7 @@ class RemoveVoltageLevelTest extends AbstractModificationTest {
     }
 
     private void addListener(Network network) {
-        network.addListener(new DefaultNetworkListener() {
+        network.addListener(new NetworkListener() {
             @Override
             public void beforeRemoval(Identifiable id) {
                 beforeRemovalObjects.add(id.getId());
