+++ iidm/iidm-tck/src/test/java/com/powsybl/iidm/network/tck/AbstractAreaTest.java
@@ -100,7 +100,7 @@ public abstract class AbstractAreaTest {
     @Test
     public void testChangesNotification() {
         // Changes listener
-        NetworkListener mockedListener = Mockito.mock(DefaultNetworkListener.class);
+        NetworkListener mockedListener = Mockito.mock(NetworkListener.class);
         // Add observer changes to current network
         network.addListener(mockedListener);
 
@@ -139,7 +139,7 @@ public abstract class AbstractAreaTest {
     @Test
     public void testSetterGetterInMultiVariants() {
         // Changes listener
-        NetworkListener mockedListener = Mockito.mock(DefaultNetworkListener.class);
+        NetworkListener mockedListener = Mockito.mock(NetworkListener.class);
         // Set observer changes
         network.addListener(mockedListener);
