+++ iidm/iidm-tck/src/test/java/com/powsybl/iidm/network/tck/AbstractMergeNetworkTest.java
@@ -622,7 +622,7 @@ public abstract class AbstractMergeNetworkTest {
     @Test
     public void testListeners() {
         MutableBoolean listenerCalled = new MutableBoolean(false);
-        NetworkListener listener = new DefaultNetworkListener() {
+        NetworkListener listener = new NetworkListener() {
             @Override
             public void onCreation(Identifiable identifiable) {
                 listenerCalled.setTrue();
