+++ iidm/iidm-tck/src/test/java/com/powsybl/iidm/network/tck/AbstractSubstationTest.java
@@ -10,7 +10,6 @@ package com.powsybl.iidm.network.tck;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.ContainerType;
 import com.powsybl.iidm.network.Country;
-import com.powsybl.iidm.network.DefaultNetworkListener;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.NetworkListener;
 import com.powsybl.iidm.network.Substation;
@@ -57,9 +56,9 @@ public abstract class AbstractSubstationTest {
         assertEquals("new tso", substation.getTso());
 
         // Create mocked network listeners
-        NetworkListener exceptionListener = mock(DefaultNetworkListener.class);
+        NetworkListener exceptionListener = mock(NetworkListener.class);
         doThrow(new UnsupportedOperationException()).when(exceptionListener).onUpdate(any(), anyString(), anyString(), any(), any());
-        NetworkListener mockedListener = mock(DefaultNetworkListener.class);
+        NetworkListener mockedListener = mock(NetworkListener.class);
         // Test without listeners registered
         substation.addGeographicalTag("no listeners");
         verifyNoMoreInteractions(mockedListener);
