+++ iidm/iidm-tck/src/test/java/com/powsybl/iidm/network/tck/AbstractTapChangerTest.java
@@ -155,7 +155,7 @@ public abstract class AbstractTapChangerTest {
         }
 
         // Changes listener
-        NetworkListener mockedListener = mock(DefaultNetworkListener.class);
+        NetworkListener mockedListener = mock(NetworkListener.class);
         // Add observer changes to current network
         network.addListener(mockedListener);
         // Changes will raise notifications
